Reortanize `Str` tests
