+++ app/Str.php
@@ -102,8 +102,6 @@ final class Str
 
 	/**
 	 * Prevent instantiation.
-	 *
-	 * @codeCoverageIgnore
 	 */
 	final private function __construct()
 	{
