+++ tests/app/StrTest.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Syntatis\Utils\Tests\Str;
+namespace Syntatis\Utils\Tests;
 
+use Error;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\TestDox;
 use PHPUnit\Framework\TestCase;
@@ -15,13 +16,23 @@ use function Syntatis\Utils\kebabcased;
 use function Syntatis\Utils\lowercased;
 use function Syntatis\Utils\macrocased;
 use function Syntatis\Utils\pascalcased;
+use function Syntatis\Utils\pluralize;
 use function Syntatis\Utils\sentencecased;
+use function Syntatis\Utils\singularize;
+use function Syntatis\Utils\slugify;
 use function Syntatis\Utils\snakecased;
 use function Syntatis\Utils\titlecased;
 use function Syntatis\Utils\uppercased;
 
-class CaseConverterTest extends TestCase
+class StrTest extends TestCase
 {
+	public function testInstance(): void
+	{
+		$this->expectException(Error::class);
+
+		new Str();
+	}
+
 	/**
 	 * @dataProvider dataCamelCased
 	 * @testdox it can convert string to camelcase
@@ -122,6 +133,44 @@ class CaseConverterTest extends TestCase
 		$this->assertEquals($expect, Str::toSentenceCase($value));
 	}
 
+	/** @dataProvider dataSingularize */
+	public function testSingularize(string $value, string $expect): void
+	{
+		$this->assertSame(Str::toSingular($value), $expect);
+		$this->assertSame(singularize($value), $expect);
+	}
+
+	public function dataSingularize(): iterable
+	{
+		yield ['tests', 'test'];
+		yield ['children', 'child'];
+		yield ['people', 'person'];
+	}
+
+	/** @dataProvider dataPluralize */
+	public function testPluralize(string $value, string $expect): void
+	{
+		$this->assertSame(Str::toPlural($value), $expect);
+		$this->assertSame(pluralize($value), $expect);
+	}
+
+	public function dataPluralize(): iterable
+	{
+		yield ['test', 'tests'];
+		yield ['child', 'children'];
+		yield ['person', 'people'];
+	}
+
+	/**
+	 * @dataProvider dataSlugify
+	 * @testdox it can transform string to a url friendly format
+	 */
+	public function testSlugify(string $value, string $expect): void
+	{
+		$this->assertSame(Str::toSlug($value), $expect);
+		$this->assertSame($expect, slugify($value));
+	}
+
 	public function dataCamelCased(): iterable
 	{
 		return [['foo_bar', 'fooBar'], ['foo-bar', 'fooBar'], ['foo bar', 'fooBar'], ['fooBar', 'fooBar'], ['FooBar', 'fooBar']];
@@ -170,5 +219,48 @@ class CaseConverterTest extends TestCase
 	public function dataSentenceCased(): iterable
 	{
 		return [['foo_bar', 'Foo bar'], ['foo-bar', 'Foo bar'], ['foo bar', 'Foo bar'], ['fooBar', 'Foo bar'], ['FooBar', 'Foo bar']];
+	}
+
+	public function dataSlugify(): iterable
+	{
+		yield [
+			'Testing_Creating a -Slug from a random-string!@#',
+			'testing-creating-a-slug-from-a-random-string',
+		];
+
+		yield [
+			'Contesta el teléfono',
+			'contesta-el-telefono',
+		];
+
+		yield [
+			'den hund füttern',
+			'den-hund-fuettern',
+		];
+
+		yield [
+			'Jsem král na severu',
+			'jsem-kral-na-severu',
+		];
+
+		yield [
+			'test1::test2',
+			'test1-test2',
+		];
+
+		yield [
+			'test1$test2',
+			'test1-test2',
+		];
+
+		yield [
+			'TESTAb-test2',
+			'testab-test2',
+		];
+
+		yield [
+			'año',
+			'ano',
+		];
 	}
 }
